arm64: mm: ensure that the zero page is visible to the page table walker
In paging_init, we allocate the zero page, memset it to zero and then
point TTBR0 to it in order to avoid speculative fetches through the
identity mapping.

In order to guarantee that the freshly zeroed page is indeed visible to
the page table walker, we need to execute a dsb instruction prior to
writing the TTBR.

Note from crpalmer: The v3.10.y version of this code would have used
different flags, but since we don't "need to drop the 'ishst'" we can
keep the better version for 3.14+ kernels.

Change-Id: I367621f84a8570e1582e065714baefa3fc58989a
Cc: <stable@vger.kernel.org> # v3.14+, for older kernels need to drop the 'ishst'
Signed-off-by: Will Deacon <will.deacon@arm.com>